reorganizing code a bit
